Import Upstream version 5.50

+++ gobex/gobex-packet.h
@@ -26,8 +26,8 @@
 #include <stdarg.h>
 #include <glib.h>
 
-#include <gobex/gobex-defs.h>
-#include <gobex/gobex-header.h>
+#include "gobex/gobex-defs.h"
+#include "gobex/gobex-header.h"
 
 /* Request opcodes */
 #define G_OBEX_OP_CONNECT			0x00
@@ -98,9 +98,9 @@ gboolean g_obex_packet_set_data(GObexPacket *pkt, const void *data, gsize len,
 						GObexDataPolicy data_policy);
 const void *g_obex_packet_get_data(GObexPacket *pkt, gsize *len);
 GObexPacket *g_obex_packet_new(guint8 opcode, gboolean final,
-						guint8 first_hdr_id, ...);
+						guint first_hdr_id, ...);
 GObexPacket *g_obex_packet_new_valist(guint8 opcode, gboolean final,
-					guint8 first_hdr_id, va_list args);
+					guint first_hdr_id, va_list args);
 void g_obex_packet_free(GObexPacket *pkt);
 
 GObexPacket *g_obex_packet_decode(const void *data, gsize len,
